bug reload and undefined were fixed
