+++ Client/src/pages/ingredients.tsx
@@ -54,7 +54,6 @@ const CreateData = (): Ingredient[] => {
   }, [rowsData]);
   
   if (!dataLoaded) {
-
     return []; 
   }
