+++ client/src/hooks/services/company.ts
@@ -1,11 +1,11 @@
 import {useMutation, useQuery} from "@apollo/react-hooks";
 import {getCompany, updateCompany} from "@/services/Company-Service";
 
-export const useGetCompanyQuery = () => {
+export const useGetCompanyQuery = ( () => {
     return useQuery(getCompany, {
         fetchPolicy: "no-cache"
     });
-}
+})
 
 export const useUpdatePhotoCompanyMutation = () => {
     return useMutation(updateCompany);
