+++ client/src/hooks/services/manager.ts
@@ -1,11 +1,12 @@
 import { useMutation, useQuery} from "@apollo/react-hooks";
 import {getManager, updateManagerInformation, updatePhotoManager} from "@/services/Manager-Service";
 
-export const useGetManagerQuery = () => {
-    return useQuery(getManager, {
+export const useGetManagerQuery = (() => {
+    return useQuery(getManager,  {
         fetchPolicy: "no-cache"
+
     });
-}
+})
 
 export const useUpdatePhotoManagerMutation = () => {
     return useMutation(updatePhotoManager)
